backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
aiofiles==23.2.1
httpx==0.25.2
python-jose==3.3.0